requirements.txt
beautifulsoup4==4.4.0
PyYAML==3.11
regex==2015.06.24
nltk==3.0.3
rdflib==4.2.0
numpy==1.9.2
click==4.0
rfc3987==1.3.4
lxml==3.4.4
requests==2.7.0
Pyro4==4.39
matplotlib==1.4.3
